+++ docs/introduction/codeexamples/variableRoutes.py
@@ -3,6 +3,6 @@ app = Klein()
 
 @app.route('/user/<username>')
 def pg_user(request, username):
-    return 'Hi %s!' % (username,)
+    return f'Hi {username}!'
 
 app.run("localhost", 8080)