+++ NEWS.rst
@@ -7,6 +7,7 @@ NEXT
  * Python 3.5 is no longer supported by Klein.
  * Python 3.9 is now supported by Klein. [`#412 <https://github.com/twisted/klein/pull/412>`_]
  * Klein now exports (incomplete, but growing) type hints. [`#379 <https://github.com/twisted/klein/pull/379>`_]
+ * ``Plating`` now sets the ``Content-Type`` header to ``application/json`` instead of ``text/json; charset=utf8``.
 
 20.6.0 - 2020-06-07
 -------------------